Fix on 32 bit Julia

+++ src/formula.jl
@@ -226,7 +226,7 @@ function rereference_formulae(ws::Worksheet, cell::Cell)
     end
 end
 
-function rereference_formulae(ws::Worksheet, oldcell::Cell, f::AbstractFormula, newrng::CellRange, newid::Int64)
+function rereference_formulae(ws::Worksheet, oldcell::Cell, f::AbstractFormula, newrng::CellRange, newid::Int)
     wb = get_workbook(ws)
     oldform = f.formula
     oldunhandled = f.unhandled
